Move Xcode class to FastlaneCore Don’t use clean if not necessary
