+++ gym/lib/gym.rb
@@ -6,7 +6,6 @@ require 'gym/runner'
 require 'gym/error_handler'
 require 'gym/options'
 require 'gym/detect_values'
-require 'gym/xcode'
 
 require 'fastlane_core'
 require 'terminal-table'
